Updated XPlane SDK to version 2.1.3

+++ xpcPlugin/SDK/CHeaders/Wrappers/XPCBroadcaster.cpp
@@ -42,7 +42,7 @@ void		XPCBroadcaster::RemoveListener(
 }				
 
 void		XPCBroadcaster::BroadcastMessage(
-				long		inMessage,
+				int			inMessage,
 				void *		inParam)
 {
 	ListenerVector::iterator iter;
