+++ xpcPlugin/SDK/CHeaders/Wrappers/XPCWidget.h
@@ -14,8 +14,8 @@ public:
 								XPCWidget *		inObject,
 								XPWidgetMessage	inMessage,
 								XPWidgetID		inWidget,
-								long			inParam1,
-								long			inParam2)=0;
+								intptr_t		inParam1,
+								intptr_t		inParam2)=0;
 								
 };
 
@@ -56,16 +56,16 @@ public:
 	virtual	int			HandleWidgetMessage(
 								XPWidgetMessage			inMessage,
 								XPWidgetID				inWidget,
-								long					inParam1,
-								long					inParam2);
+								intptr_t				inParam1,
+								intptr_t				inParam2);
 		
 private:
 
 	static	int			WidgetCallback(
 								XPWidgetMessage			inMessage,
 								XPWidgetID				inWidget,
-								long					inParam1,
-								long					inParam2);
+								intptr_t				inParam1,
+								intptr_t				inParam2);
 		
 	typedef	std::pair<XPCWidgetAttachment *, bool>	AttachmentInfo;
 	typedef	std::vector<AttachmentInfo>				AttachmentVector;
